✨Feat: user-model 작성
