+++ app/models/base_user_models.py
@@ -50,16 +50,3 @@ class BaseUser(models.Model):
 
     class Meta:
         table = "BASE_USER"
-
-
-class UserBan(models.Model):
-    id = fields.IntField(pk=True)
-    user = fields.ForeignKeyField("app.models.BaseUser", related_name="bans")
-    reason = fields.TextField(null=True)
-    banned_at = fields.DatetimeField(auto_now_add=True)
-    banned_until = fields.DatetimeField(null=True)
-    banned_by = fields.ForeignKeyField("app.models.BaseUser", related_name="bans_given")
-    email_sent = fields.BooleanField(default=False)
-
-    class Meta:
-        table = "USER_BAN"
